Add GCP authentication support (#1)
* Add GCP authentication support
