+++ pkg/authenticator/k8s/testdata/example.csr
