+++ pkg/config/config.go
@@ -1,0 +1,9 @@
+package config
+
+import "time"
+
+// Config global configs used by the authenticator
+type Config struct {
+	TokenRefreshTimeout time.Duration
+	ContainerMode       string
+}
