+++ pkg/log/log_messages.go
@@ -60,7 +60,7 @@ const CAKC044 string = "CAKC044 Buffer time:  %v"
 const CAKC045 string = "CAKC045 Login request to: %s"
 const CAKC046 string = "CAKC046 Authn request to: %s"
 const CAKC047 string = "CAKC047 Waiting for %s to re-authenticate"
-const CAKC048 string = "CAKC048 Kubernetes Authenticator Client v%s starting up..."
+const CAKC048 string = "CAKC048 Conjur Authenticator Client v%s starting up..."
 const CAKC049 string = "CAKC049 Loaded client certificate successfully from %s"
 const CAKC050 string = "CAKC050 Deleted client certificate from memory"
 const CAKC051 string = "CAKC051 Waiting for file %s to become available..."
@@ -70,3 +70,6 @@ const CAKC054 string = "CAKC054 Failed to delete file %s"
 const CAKC055 string = "CAKC055 Cert placement failed with the following error:\n%s"
 const CAKC056 string = "CAKC056 File %s does not exist"
 const CAKC057 string = "CAKC057 Removing file %s"
+const CAKC058 string = "CAKC058 Initializing k8s authenticator"
+const CAKC059 string = "CAKC059 Initializing gcp authenticator"
+const CAKC060 string = "CAKC060 Invalid conjur authentication type '%s'"
